adjusted a few things

+++ src/emulator/AddressMode.h
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-uint16_t immediate_address_mode(CPU &cpu);
+uint8_t immediate_address_mode(CPU &cpu);
 
 uint16_t zero_page_address_mode(CPU &cpu);
 uint16_t zero_page_address_mode_X(CPU &cpu);
